pyproject.toml
[project]
name = "fastmcp-implementation"
version = "0.1.0"
description = "Production-ready MCP server and client using FastMCP with Python 3.13"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "MCP Team", email = "team@example.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "client"]
